Move monetized attributes tests to a new spec file for better organization
